// src/generated/prisma/enums.ts
export const RecordStatus = {
  active: 'active',
  inactive: 'inactive',
  blocked: 'blocked',
  archived: 'archived'
} as const

export type RecordStatus = (typeof RecordStatus)[keyof typeof RecordStatus]


export const UserRole = {
  user: 'user',
  admin: 'admin',
  super_admin: 'super_admin'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AccountType = {
  custom: 'custom',
  google: 'google'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]